+++ eth_defi/enzyme/deployment.py
@@ -616,4 +616,3 @@ class EnzymeDeployment:
             usdc=usdc,
         )
 
-]